consolidate mountpoint validation a bit
